mundança de int para long long int que é o ideal para armazenar cpf

+++ EDA(trab_4)/Insere.h
@@ -1,7 +1,7 @@
 #ifndef INSERE
 #define INSERE
-void enderecamento_interno_v2(int* tabela, int val, int tam);
-void enderecamento_interno_v1(int* tabela, int val, int tam);
+void enderecamento_interno_v2(long long int* tabela,long long int val, int tam);
+void enderecamento_interno_v1(long long int* tabela,long long int val, int tam);
 #endif // !INSERE
