+++ EDA(trab_4)/cria_e_inicializa_tabela.h
@@ -2,12 +2,12 @@
 #define CRIA_TABELA
 
 struct node {
-    int chave;
+   long long int chave;
     struct node* prox;
 };
 typedef struct node Node;
 
-int* variacao_interior(const int tam);
+long long int* variacao_interior(const int tam);
 
 Node** variacao_exterior(int tam);
