+++ EDA(trab_4)/metodo_divisao.h
@@ -1,7 +1,7 @@
 #ifndef METODO_DIVISAO
 
-int metodo_divisao_v1(int chave, int tam);
+int metodo_divisao_v1(long int chave, int tam);
 
-int metodo_divisao_v2(int chave, int tam, int c);
+int metodo_divisao_v2(long int chave, int tam, int c);
 
 #endif // ! METODO_DIVISAO
